Feedback for jest-runtime CLI
